substack/articles/2025.10.29-airflow_duckdb/pyproject.toml
[project]
name = "2025-10-29-airflow-duckdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow==3.1.1",
    "duckdb>=1.4.1",
]
